+++ vitest.config.ts
@@ -72,6 +72,7 @@ export default defineConfig({
       '@base-ui/react/form',
       '@base-ui/react/menu',
       '@base-ui/react/menubar',
+      '@base-ui/react/meter',
       '@base-ui/react/navigation-menu',
       '@base-ui/react/number-field',
       '@base-ui/react/otp-field',